+++ scout/constants/filters.py
@@ -7,6 +7,7 @@ CLINICAL_FILTER_BASE = {
     "functional_annotations": SEVERE_SO_TERMS,
     "clinsig": [4, 5],
     "clinsig_confident_always_returned": True,
+    "prioritise_clinvar": True,
 }
 
 CLINICAL_FILTER_BASE_SV = {